fix: TSconfig to allow file structure for surveys (#1001)
* Refactor

* Fix

* fix?

* fix?

* option 2

* refactor and fixes

---------
